Refactor gift card
