+++ example/themes/original/src/components/GiftCard/GiftCardUI/styles.tsx
