+++ example/themes/original/src/components/GiftCard/SendGiftCard/index.tsx
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react'
 import { StyleSheet, Platform, Alert } from 'react-native';
 import { useLanguage, SendGiftCard as SendGiftCardController  } from 'ordering-components/native';
 import { useTheme } from 'styled-components/native';
-import { OText, OButton, OInput } from '../shared';
+import { OText, OButton, OInput } from '../../shared';
 import { useForm, Controller } from 'react-hook-form'
 
 import {
